requirements.txt
pytube
moviepy
python-dotenv
google-cloud-speech